Fix MEDIUM security issues from SSH jail + general audits
- Webhook SSRF: reject private/loopback IPs in NOTIFY_WEBHOOK
- Nginx config injection: sanitize bot rule patterns (strip ', ;, {, })
- Health endpoint: return minimal status without auth, details require key
- Config permissions: warn if world-readable with API_KEY set
- Panel: cast $port to (int) in HtmlString
- Nginx.conf writes: atomic temp+rename pattern
- API key install: generate+write in Python (not exposed in ps)
- sshd: add LoginGraceTime 60 + MaxStartups 10:30:60

+++ panel/Pages/Security.php
@@ -439,7 +439,7 @@ class Security extends Page implements HasActions, HasForms
     {
         $s = $this->client()->get('/ssh/sshd-settings') ?? [];
         $passAuth = $s['password_auth'] ?? true;
-        $port = $s['port'] ?? 22;
+        $port = (int) ($s['port'] ?? 22);
 
         $jailEnabled = in_array($this->moduleStates['SSHJAIL_ENABLED'] ?? false, [true, 'yes', '1', 1]);
